+++ local/components/b24tech/schedule.payments/class.php
@@ -12,12 +12,14 @@ Loader::includeModule("sale");
 class SchedulePayments extends CBitrixComponent
 {
     const HL = 3;
+    const HL_LOG = 6;
     const HL_PERIOD = 5;
     const PLATFORM_ID = 143;
     const TARGET_ID = 145;
 
     public $LAST_ERROR;
     public $GROUP_ID;
+    public $PERIOD_ID;
     public $loadData = false;
 
     protected function prepareData()
@@ -369,7 +371,8 @@ class SchedulePayments extends CBitrixComponent
         if ($arPeriodFields['UF_DEFAULT'] == 1) {
             $arPeriodFields['UF_DEFAULT'] = 0;
         }
-        $arPeriodFields['UF_STATUS'] = 'C'; // clear status
+        $arPeriodFields['UF_STATUS'] = ''; // clear status
+        $arPeriodFields['UF_IS_COPY'] = 'Y';
         unset($arPeriodFields['ID']);
         $entityPeriod = GetEntityDataClass(self::HL_PERIOD);
         $entity = GetEntityDataClass(self::HL);
@@ -492,9 +495,10 @@ class SchedulePayments extends CBitrixComponent
                             foreach ($arItems as $arItem) {
                                 $entity::update($arItem['ID'], array(
                                         'UF_PAID' => 0,
-                                        'UF_SPEND_NDS' => 0
+                                        //'UF_SPEND_NDS' => 0
                                     )
                                 );
+                                $this->calculate($arItem['ID']);
                             }
                         }
                     }
@@ -668,7 +672,11 @@ class SchedulePayments extends CBitrixComponent
                                     }
                                 }
                                 if ($error === false) {
+
+                                    $this->log($key, $arFields);
+
                                     $result = $entity::update($key, $arFields);
+
                                     $this->calculate($key);
                                     if (!$result->isSuccess()) {
                                         $arResponse['result'] = false;
@@ -729,5 +737,49 @@ class SchedulePayments extends CBitrixComponent
             $this->prepareData();
             $this->includeComponentTemplate();
         }
+    }
+    public function getTableCurrntFields($ID)
+    {
+        return true;
+    }
+    public function log($id, $arFields = [])
+    {
+        $arResult = [];
+        $entityRecord = GetEntityDataClass(self::HL);
+        $entityLog = GetEntityDataClass(self::HL_LOG);
+        $ar = $entityRecord::getList(
+            [
+                'filter' => [
+                    'ID' => $id,
+                ]
+            ]
+        )->fetch();
+
+        foreach ($arFields as $code => $value) {
+            switch ($code) {
+                case 'UF_MUST_PAY':
+                    $arResult = 'Поле "Должны оплатить" изменено с ' . $ar[$code] . ' на ' . $value;
+                    break;
+                case 'UF_PAID':
+                    $arResult = 'Поле "Оплачено" изменено с ' . $ar[$code] . ' на ' . $value;
+                    break;
+                case 'UF_CREDITMONEY':
+                    $arResult = 'Поле "Кредит" изменено с ' . $ar[$code] . ' на ' . $value;
+                    break;
+            }
+        }
+
+
+        global $USER;
+        $arLogFields = [
+            'UF_GROUP_ID' => $this->arResult['ID'],
+            'UF_PERIOD_ID' => $ar['UF_PERIOD'],
+            'UF_PLATFORM' => $ar['UF_PLATFORM'],
+            'UF_TARGET' => $ar['UF_TARGET'],
+            'UF_USER_ID' => $USER->GetID(),
+            'UF_RECORD' => $id,
+            'UF_HISTORY' => $arResult
+        ];
+        $entityLog::add($arLogFields);
     }
 }